Update instructions in README, add comments in draco.pl, fix bug
$comment in print_comment_chain() was changed to $comment_data to
prevent confusion in future.

We have to check if url has "/" at the end because if it does have it
then previously we were appending "$comment_data->{id}.json" to it
which makes the url look like:

      http://<reddit>/<post_id>//<comment_id>.json

And reddit doesn't like that, reddit wants it to be:

      http://<reddit>/<post_id>/<comment_id>.json

Notice the double '/' in previous link.